feat: separate income component to other file
