+++ src/components/IncomeItem/index.js
@@ -1,0 +1,1 @@
+export { IncomeItem } from './IncomeItem';
